Touch Readme to prevent automatic deploys
There's a bug in the jenkins pipeline this release uses. If we ever need to wipe jenkins and start again, this release would be automatically deployed because the tip of the release branch point to the same commit at the release tag. 

To fix this all we need is for there to be another commit on top of the tagged commit (this is already the case for the previous release tag).