Generate docs in build dir (#81)
Currently we generate our docs directly in the `docs` folder. However
this makes it easy to accidentally commit our _config.yml file (which is
now generated by a gradle task).

Here we switch it up and do our docs generation in the build dir, so it
can be cleaned up and so there's no more .gitignore to worry about.